[update] create initial seeds :)
