+++ data/seedHelpers/createGroupMembersSeed.js
@@ -1,0 +1,15 @@
+const groupMembers = [
+  {
+    id: 1,
+    user_id: 1,
+    user_type: 'admin',
+    group_id: 1,
+  },
+  {
+    id: 2,
+    user_id: 1,
+    user_type: 'admin',
+    group_id: 2,
+  },
+]
+module.exports = groupMembers
