+++ data/seedHelpers/createPostsSeeds.js
@@ -1,0 +1,10 @@
+const posts = [
+  {
+    id: 1,
+    user_id: 1,
+    group_id: 1,
+    content: 'This is a post',
+    title: 'YESTERDAYS DISAPPOINTMENT ....',
+  },
+]
+module.exports = posts
