+++ data/seedHelpers/createSportsSeeds.js
@@ -1,0 +1,9 @@
+const sports = [
+  { id: 1, name: 'soccer' },
+  { id: 2, name: 'football' },
+  { id: 3, name: 'hockey' },
+  { id: 4, name: 'tennis' },
+  { id: 5, name: 'basketball' },
+  { id: 6, name: 'baseball' },
+]
+module.exports = sports
