+++ data/seedHelpers/index.js
@@ -1,0 +1,9 @@
+module.exports = {
+  createAllegianceMembersSeeds: require('./createAllegianceMembersSeeds'),
+  createLeaguesSeeds: require('./createLeaguesSeeds'),
+  createPostsSeeds: require('./createPostsSeeds'),
+  createSportsSeeds: require('./createSportsSeeds'),
+  createAllegiancesSeeds: require('./createAllegianceSeeds'),
+  createUsersSeeds: require('./createUsersSeeds'),
+  createGroupsSeeds: require('./createGroupSeeds'),
+}
